requirements.txt
Django==3.2.9
djangorestframework==3.12.4
# django-cors-headers==3.10.0
dj-database-url==0.5.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
# whitenoise==6.0.0
